Updated ConfigHandler with new config.ini format and functions
Changed DEFAULT_CONFIG sections and options. This includes using dict values as options for Input Devices and Triggers sections. Example of how to use them is in test_config.py.
The functions labelled with TODO in ConfigHandler may not be needed anymore.

The following scripts are affected and need to be amended:
- color_picker_page.py
- keyword_page.py
- main.py
- paddle_ocr.py
- TTS.py