Updated config_handler.py with some arg type conversions
Added type conversions for 'usb_alt_name', 'custom_name' and 'condition' arguments in affected functions to force-convert them to String. This is to handle cases where the affected functions use those args as String-type but non-String value is passed in (e.g. integer). This update does not impact the existing calling of ConfigHandler functions. Updated functions are:
-set_cfg_input_device()
-get_cfg_input_devices()
-add_input_device()
-del_input_device()