Updated config_handler.py to handle setting of condition name
Added a new "condition_name" triggers sub-value to DEFAULT_CONFIG template. Since there is a change to DEFAULT_CONFIG template, a new config.ini will be created automatically with this new template. This new "condition_name" is to allow users to set different names to the different conditions, whereas the existing "condition" trigger sub-key serves as more of a unique ID.
The set_cfg_input_device() function has been modified with the following:
- It now takes in a new "condition_name" optional kwarg, which accepts a String value.
- Duplicate "condition_name" is rejected if it already exists in one of the conditions for that Input Device.
- Fixed a bug with the initial For.. loop that cycles through the kwargs passed in.
- Example to change condition name: set_cfg_input_device(usb_alt_name = "abcdef123", condition = "cond1", condition_name = "this is a new condition name", ...<plus other optional kwargs like tts_text = "hello", etc. if you want to set them too>)