Update cam_setup_page.py
Resolved #18 #22 
Edited the default video label height from 300 to 20. This will display the buttons and labels when input device is not connected.

When input device is connected, the connection status will change to Connected and the monitoring device button's state will be normal to allow user to toggle if they want to monitor the device or not.