Updated paddle_ocr.py to fix discoloured OCR'd alert screenshots
Removed convert_frame() as the BGR to RGB conversion was not needed when using imageio's frames for OCR since it was already in RGB format.
Removed some debugging printouts.
Increased persistence of OCR'd alert screenshots from 5min to 60min as RAM usage looks healthy. This may be increased further in the future after more testing is done with higher limits.