Update TTS.py for compatibility with portable Python distributions (like WinPython)
In portable Python distributions like WinPython, the COM library may not be auto-initialised when the script is run. Because pyttsx3 uses Windows Speech API via COM objects, if the COM library is not initialised before pyttsx3 is initialised, an error will occur. This update manually initialises the COM library at the start of the script to prevent such an error.